+++ zebra-chain/src/block.rs
@@ -1,7 +1,6 @@
 //! Definitions of block datastructures.
 #![allow(clippy::unit_arg)]
 
-mod difficulty;
 mod hash;
 mod header;
 mod height;
