+++ zebra-chain/src/work.rs
@@ -1,0 +1,7 @@
+//! Proof-of-work implementation.
+
+pub mod difficulty;
+pub mod equihash;
+
+#[cfg(test)]
+mod tests;
