+++ zebra-chain/src/work/tests.rs
@@ -1,0 +1,3 @@
+mod arbitrary;
+mod prop;
+mod vectors;
